Add moderation, reports DTOs, and profiles

+++ src/main/java/com/reddot/api/dto/VoteRequest.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class VoteRequest {
-    private short value; // +1 ou -1
+    private int value; // +1 ou -1
 }
